pyproject.toml
[project]
name = "langchain-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "langgraph>=0.2.0",
    "python-dotenv>=1.0.0",
    "pypdf>=5.1.0",
    "langchain-text-splitters>=0.3.0",
    "faiss-cpu>=1.7.4",
]
